refactor: move function `htmlToElement` to helpers
- we will use this in the stories too
