+++ src/components/helpers/helpers.ts
@@ -128,3 +128,9 @@ export function scrollIntoView(element: HTMLElement, scrollParent: HTMLElement):
     scrollParent.scrollBy(0, bottomDistance);
   }
 }
+
+export function htmlToElement(html: string): Element[] {
+  const template = document.createElement('template');
+  template.innerHTML = html.trim();
+  return Array.from(template.content.children);
+}
