+++ src/components/helpers/jest.ts
@@ -18,12 +18,6 @@ export const simulateKey = (key: string, target: Element): void => {
 export const simulateClick = (target: Element, point: ClientPoint | undefined = undefined): void =>
   simulateMouse('click', target, point);
 
-export function htmlToElement(html: string): Element[] {
-  const template = document.createElement('template');
-  template.innerHTML = html.trim();
-  return Array.from(template.content.children);
-}
-
 export async function propReflectsInAttributes(page: SpecPage, props: Record<string, unknown>): Promise<boolean> {
   // Use this helper to test if all supported props are reflected in attributes
   for (const [key, value] of Object.entries(props)) {
